+++ Enums/AgentInterfaceVfs.cs
@@ -3,5 +3,8 @@ namespace HaselTweaks.Enums;
 public enum AgentInterfaceVfs
 {
     ReceiveEvent = 0,
+    Show = 3,
+    Hide = 4,
+    IsAgentActive = 5,
     Update = 6,
 }