+++ Tweaks/EnhancedMaterialList.cs
@@ -17,6 +17,7 @@ using HaselTweaks.Enums;
 using HaselTweaks.Structs;
 using HaselTweaks.Utils;
 using Lumina.Excel.GeneratedSheets;
+using Character = FFXIVClientStructs.FFXIV.Client.Game.Character.Character;
 
 namespace HaselTweaks.Tweaks;